Fixed ! marker being summoned even if entity (Pig) has no dialogs

+++ src/main/java/dev/zanckor/mod/common/util/MCUtil.java
@@ -6,7 +6,6 @@ import dev.zanckor.api.filemanager.quest.codec.user.UserGoal;
 import dev.zanckor.api.filemanager.quest.codec.user.UserQuest;
 import dev.zanckor.example.common.enumregistry.EnumRegistry;
 import dev.zanckor.mod.QuestApiMain;
-import dev.zanckor.mod.common.network.SendQuestPacket;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.Mth;
@@ -98,7 +97,7 @@ public class MCUtil {
         float ySinRotation = yRotSin * xCosDegrees;
 
         //Distance in blocks, multiplier is applied to player reach distance
-        double viewDistance = player.getReachDistance() * multiplier;
+        double viewDistance = player.getBlockReach() * multiplier;
 
         Vec3 lookingVector = eyePos.add((double) ySinRotation * viewDistance, (double) xSinDegrees * viewDistance, (double) yCosRotation * viewDistance);
         return level.clip(new ClipContext(eyePos, lookingVector, ClipContext.Block.OUTLINE, ClipContext.Fluid.NONE, player));
@@ -193,7 +192,10 @@ public class MCUtil {
 
     public static void moveFileToUncompletedFolder(Path uncompletedQuestFolder, File file, UserQuest userQuest, Enum goalEnum) throws IOException {
         Path uncompletedPath = Paths.get(uncompletedQuestFolder.toString(), file.getName());
-        Files.move(file.toPath(), uncompletedPath);
+
+        if (file.exists()) {
+            Files.move(file.toPath(), uncompletedPath);
+        }
         LocateHash.movePathQuest(userQuest.getId(), uncompletedPath, goalEnum);
     }
